apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
alpaca-py==0.33.1
boto3==1.36.4
pandas-ta==0.4.71b0
yfinance==0.2.51
pydantic-settings==2.7.1
python-dotenv==1.0.1
apscheduler==3.11.0
httpx==0.28.1